+++ HMM1/src/types.jl
@@ -3,9 +3,9 @@ module Types
 export Probability, A, B, HMM, StochasticVector, ObservationSpace
 
 struct Probability
-    value::Float32
+    value::Float64
 
-    function Probability(value::Float32)
+    function Probability(value::Float64)
         # Überprüfe, ob die Matrix quadratisch ist
         if (value < 0.) || (value > 1.)
             error("Probability $value not in [0,1].")
@@ -17,7 +17,7 @@ end
 
 struct A
     dimension::UInt
-    transitionMatrix::Array{Float32, 2}
+    transitionMatrix::Array{Float64, 2}
     
     function A(dim, transMatrix)
         a,b = size(transMatrix)
@@ -32,7 +32,7 @@ end
 
 struct B
     dimension::Tuple{Int,Int}
-    transitionMatrix::Array{Float32, 2}
+    transitionMatrix::Array{Float64, 2}
 
     function B((dim1, dim2), transMatrix)
         a,b = size(transMatrix)
@@ -71,7 +71,7 @@ struct ObservationSpace
 end
 
 struct StochasticVector
-    probabilities::Vector{Float32}
+    probabilities::Vector{Float64}
 end
